requirements.txt
flask
gunicorn
openai
requests
python-dotenv
faiss-cpu
llama-index
langchain
certifi
datasets
elevenlabs
sentence-transformers
ipykernel
pydub
numpy